+++ src/stam_annotator/opf_loader.py
@@ -3,8 +3,7 @@ from typing import Dict
 from pydantic import BaseModel, Field, ValidationInfo, field_validator
 
 from stam_annotator.definations import OPF_DIR
-from stam_annotator.load_yaml import load_opf_annotations_from_yaml
-from stam_annotator.utility import get_uuid
+from stam_annotator.utility import get_uuid, load_opf_annotations_from_yaml
 
 
 class Span(BaseModel):
